Patch Readline module so we can use start/finish indices in readline completion
The line buffer is not always a reliable source of truth because it is
not updated synchronously (which is perfectly understandable).

Hence when you type:

    > help[ENTER]
    ...
    > h

The buffer can contain the previous string "help\n", so you cannot guess
the current state of the current line from it.

It seems it's updated every 3 or 4 characters or every few seconds so
it's still a valuable source of information for the day we will start
completing command arguments..