Ignore Ctrl+C in current shell
Note that it breaks when the current line buffer is not empty (its
content is left behind and never erased, and basically Crystal bindings
don't expose methods to get rid of that ; not only Crystal's fault
though because at least on OSX, common functions to erase readline's
line buffer are marked as "not implemented". Ouch.